Refactor types into individual model files
- Moved all types from `types.ts` to separate files under `models/` directory
- Updated services to use the new model imports
- Improved project structure for better scalability and maintainability
