+++ frontend-react/src/models/Change.ts
@@ -1,0 +1,8 @@
+import {PlayerInPosition} from "./PlayerInPosition.ts";
+
+export interface Change {
+    id: string;
+    playerOut: PlayerInPosition;
+    playerIn: PlayerInPosition;
+    time: Date;
+}
