+++ frontend-react/src/models/Player.ts
@@ -1,0 +1,4 @@
+export interface Player {
+    id: string;
+    name: string;
+}
