+++ frontend-react/src/models/PlayerInPosition.ts
@@ -1,0 +1,7 @@
+import {Player} from "./Player.ts";
+
+export interface PlayerInPosition {
+    player: Player;
+    position: string;
+    team: string;
+}
