+++ frontend-react/src/models/Season.ts
@@ -1,0 +1,3 @@
+export interface Season {
+    id: string;
+}
