+++ frontend-react/src/models/Serie.ts
@@ -1,0 +1,7 @@
+export interface Serie {
+    id: string;
+    idSeason: string;
+    name: string;
+    startDate: Date;
+    endDate: Date;
+}
