+++ frontend-react/src/models/Team.ts
@@ -1,0 +1,9 @@
+export interface Team {
+    id: string;
+    name: string;
+    initials: string;
+    color: string;
+    winGames: number;
+    loseGames: number;
+    totalRuns: number;
+}
