+++ frontend-react/src/services/config/apiClient.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { API_URL } from './config';
+import { API_URL } from './config.ts';
 
 const apiClient = axios.create({
     baseURL: API_URL,
