+++ frontend-react/src/services/users/all/serieService.ts
@@ -1,6 +1,8 @@
-import apiClient from './apiClient';
-import { Game, Team } from './types';
-import { Series } from './types';
+import apiClient from '../../config/apiClient.ts';
+import {Serie} from "../../../models/Serie.ts";
+import {Team} from "../../../models/Team.ts";
+import {Game} from "../../../models/Game.ts";
+
 
 const SERIES_URL = '/series';
 
@@ -17,7 +19,7 @@ export const fetchTeamsInSeries = async (seasonId: string, serieId: string): Pro
 };
 
 // Fetch all series
-export const fetchAllSeries = async (): Promise<Series[]> => {
+export const fetchAllSeries = async (): Promise<Serie[]> => {
     const response = await apiClient.get(SERIES_URL);
     return response.data;
 };
